backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
requests==2.31.0
selenium==4.15.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
celery==5.3.4
redis==4.5.5
django-cors-headers==4.3.1